WIP: Adding speculate / bake-scripts
