+++ bake-scripts/config.json
@@ -1,0 +1,11 @@
+{
+    "name": "my-service-name",
+    "environment": "int",
+    "release": "0.1.0.841-1.bbc.el7.noarch",
+    "configuration": {
+        "API_URL": "https://digital-paper-edit-api.int.tools.bbc.co.uk"
+    },
+    "secure_configuration": {
+        "password": "sekrit"
+    }
+}
