+++ public/env.js
@@ -1,0 +1,1 @@
+window.env = { API_URL: 'http://localhost:8080' };
