+++ server.js
@@ -1,18 +1,41 @@
-const os = require('os');
-const cluster = require('cluster');
+const path = require('path');
+const express = require('express');
 
-const cpuCount = os.cpus().length;
+const ENV = process.env.NODE_ENV || 'dev';
+const PORT = process.env.PORT || 3000;
 
-if (cluster.isMaster) {
-  for (let i = 0; i < cpuCount; i++) {
-    cluster.fork();
-  }
+const app = express();
 
-  cluster.on('exit', (worker) => {
-    stats.increment('processes_killed');
-    console.log(`Worker ${ worker.process.pid } died`);
-    cluster.fork();
+const staticIndex = path.join(
+  __dirname,
+  '/build',
+);
+
+app.use(express.static(staticIndex));
+
+app.use((err, req, res, next) => {
+  const statusCode = err.statusCode || 500;
+
+  res.status(statusCode).json({
+    status: statusCode,
+    message: err.message,
   });
-} else {
-  require('./app');
-}
+});
+
+const server = app.listen(PORT, () => {
+  console.log(`ENV: ${ ENV } | listening on port ${ PORT }`);
+});
+
+server.on('error', (err) => {
+  console.error(err);
+});
+
+app.get('/status', (req, res) => {
+  res.sendStatus(200);
+});
+
+app.get('/', (req, res) => {
+  res.sendFile(path.join(staticIndex, '/index.html'));
+});
+
+module.exports = app;
